Move stage to impl package
